pyproject.toml
[project]
name = "medical-multi-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.1",
    "langchain-openai>=1.2.1",
    "langgraph>=1.1.9",
    "mcp>=1.27.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "streamlit>=1.56.0",
    "uvicorn>=0.46.0",
]
